Fix RQ wrongly moving jobs to FailedJobRegistry
Something changed in python-rq and the old code was behaving in a way
that if a job ran for longer than 2 min it would be automatically set as
failed, but it would continue running.

This causes a problem in the UI because it is as if the job stopped, but
it actually didn't.

https://github.com/getredash/redash/pull/7186